+++ examples/workflow-provision-orders/main.ts
@@ -1,88 +1,22 @@
-import { createMachine, fromPromise, interpret } from 'xstate';
+import { fromPromise, createActor, setup } from 'xstate';
+
+interface Order {
+  id: string;
+  item: string;
+  quantity: string;
+}
 
 // https://github.com/serverlessworkflow/specification/tree/main/examples#provision-orders-example
-export const workflow = createMachine(
-  {
-    id: 'provisionorders',
-    types: {} as {
-      context: {
-        order: {
-          id: string;
-          item: string;
-          quantity: string;
-        };
-      };
+export const workflow = setup({
+  types: {
+    context: {} as {
+      order: Order;
     },
-    initial: 'ProvisionOrder',
-    context: ({ input }) => ({
-      order: input.order
-    }),
-    states: {
-      ProvisionOrder: {
-        invoke: {
-          src: 'provisionOrderFunction',
-          input: ({ context }) => ({
-            order: context.order
-          }),
-          onDone: 'ApplyOrder',
-          onError: [
-            {
-              guard: ({ event }) => event.data.message === 'Missing order id',
-              target: 'Exception.MissingId'
-            },
-            {
-              guard: ({ event }) => event.data.message === 'Missing order item',
-              target: 'Exception.MissingItem'
-            },
-            {
-              guard: ({ event }) =>
-                event.data.message === 'Missing order quantity',
-              target: 'Exception.MissingQuantity'
-            }
-          ]
-        }
-      },
-      ApplyOrder: {
-        invoke: {
-          src: 'applyOrderWorkflowId',
-          onDone: 'End'
-        }
-      },
-      End: {
-        type: 'final'
-      },
-      Exception: {
-        initial: 'MissingId',
-        states: {
-          MissingId: {
-            invoke: {
-              src: 'handleMissingIdExceptionWorkflow',
-              onDone: 'End'
-            }
-          },
-          MissingItem: {
-            invoke: {
-              src: 'handleMissingItemExceptionWorkflow',
-              onDone: 'End'
-            }
-          },
-          MissingQuantity: {
-            invoke: {
-              src: 'handleMissingQuantityExceptionWorkflow',
-              onDone: 'End'
-            }
-          },
-          End: {
-            type: 'final'
-          }
-        },
-        onDone: 'End'
-      }
-    }
+    input: {} as { order: Order }
   },
-  {
-    actors: {
-      provisionOrderFunction: fromPromise(async ({ input }) => {
+  actors: {
+    provisionOrderFunction: fromPromise(
+      async ({ input }: { input: { order: Order } }) => {
         console.log('starting provisionOrderFunction');
         await new Promise((resolve) => setTimeout(resolve, 1000));
         if (!input.order.id) {
@@ -98,36 +32,107 @@ export const workflow = createMachine(
         return {
           order: input.order
         };
-      }),
-      applyOrderWorkflowId: fromPromise(async () => {
-        console.log('starting applyOrderWorkflowId');
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-        console.log('finished applyOrderWorkflowId');
-        return;
-      }),
-      handleMissingIdExceptionWorkflow: fromPromise(async () => {
-        console.log('starting handleMissingIdExceptionWorkflow');
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-        console.log('finished handleMissingIdExceptionWorkflow');
-        return;
-      }),
-      handleMissingItemExceptionWorkflow: fromPromise(async () => {
-        console.log('starting handleMissingItemExceptionWorkflow');
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-        console.log('finished handleMissingItemExceptionWorkflow');
-        return;
-      }),
-      handleMissingQuantityExceptionWorkflow: fromPromise(async () => {
-        console.log('starting handleMissingQuantityExceptionWorkflow');
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-        console.log('finished handleMissingQuantityExceptionWorkflow');
-        return;
-      })
+      }
+    ),
+    applyOrderWorkflowId: fromPromise(async () => {
+      console.log('starting applyOrderWorkflowId');
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      console.log('finished applyOrderWorkflowId');
+      return;
+    }),
+    handleMissingIdExceptionWorkflow: fromPromise(async () => {
+      console.log('starting handleMissingIdExceptionWorkflow');
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      console.log('finished handleMissingIdExceptionWorkflow');
+      return;
+    }),
+    handleMissingItemExceptionWorkflow: fromPromise(async () => {
+      console.log('starting handleMissingItemExceptionWorkflow');
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      console.log('finished handleMissingItemExceptionWorkflow');
+      return;
+    }),
+    handleMissingQuantityExceptionWorkflow: fromPromise(async () => {
+      console.log('starting handleMissingQuantityExceptionWorkflow');
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      console.log('finished handleMissingQuantityExceptionWorkflow');
+      return;
+    })
+  }
+}).createMachine({
+  id: 'provisionorders',
+
+  initial: 'ProvisionOrder',
+  context: ({ input }) => ({
+    order: input.order
+  }),
+  states: {
+    ProvisionOrder: {
+      invoke: {
+        src: 'provisionOrderFunction',
+        input: ({ context }) => ({
+          order: context.order
+        }),
+        onDone: 'ApplyOrder',
+        onError: [
+          {
+            guard: ({ event }) =>
+              (event.error as any).message === 'Missing order id',
+            target: 'Exception.MissingId'
+          },
+          {
+            guard: ({ event }) =>
+              (event.error as any).message === 'Missing order item',
+            target: 'Exception.MissingItem'
+          },
+          {
+            guard: ({ event }) =>
+              (event.error as any).message === 'Missing order quantity',
+            target: 'Exception.MissingQuantity'
+          }
+        ]
+      }
+    },
+    ApplyOrder: {
+      invoke: {
+        src: 'applyOrderWorkflowId',
+        onDone: 'End'
+      }
+    },
+    End: {
+      type: 'final'
+    },
+    Exception: {
+      initial: 'MissingId',
+      states: {
+        MissingId: {
+          invoke: {
+            src: 'handleMissingIdExceptionWorkflow',
+            onDone: 'End'
+          }
+        },
+        MissingItem: {
+          invoke: {
+            src: 'handleMissingItemExceptionWorkflow',
+            onDone: 'End'
+          }
+        },
+        MissingQuantity: {
+          invoke: {
+            src: 'handleMissingQuantityExceptionWorkflow',
+            onDone: 'End'
+          }
+        },
+        End: {
+          type: 'final'
+        }
+      },
+      onDone: 'End'
     }
   }
-);
+});
 
-const actor = interpret(workflow, {
+const actor = createActor(workflow, {
   input: {
     order: {
       id: '',