+++ packages/core/src/State.ts
@@ -54,6 +54,9 @@ interface MachineSnapshotBase<
   TOutput,
   _TUnusedButLeftForCompatReasons = never
 > {
+  /**
+   * The state machine that produced this state snapshot.
+   */
   machine: StateMachine<
     TContext,
     TEvent,
@@ -67,8 +70,32 @@ interface MachineSnapshotBase<
     unknown,
     TOutput
   >;
+  /**
+   * The tags of the active state nodes that represent the current state value.
+   */
   tags: Set<string>;
+  /**
+   * The current state value.
+   *
+   * This represents the active state nodes in the state machine.
+   * - For atomic state nodes, it is a string. 
+   * - For compound parent state nodes, it is an object where:
+   *   - The key is the parent state node's key
+   *   - The value is the current state value of the active child state node(s)
+   * 
+   * @example
+  ```ts
+  // single-level state node
+  snapshot.value; // => 'yellow'
+
+  // nested state nodes
+  snapshot.value; // => { red: 'wait' }
+  ```
+   */
   value: TStateValue;
+  /**
+   * The current status of this snapshot.
+   */
   status: 'active' | 'done' | 'error' | 'stopped';
   error: unknown;
   context: TContext;
@@ -84,8 +111,8 @@ interface MachineSnapshotBase<
   children: TChildren;
 
   /**
-   * Whether the current state value is a subset of the given parent state value.
-   * @param testValue
+   * Whether the current state value is a subset of the given partial state value.
+   * @param partialStateValue
    */
   matches: (
     this: MachineSnapshot<
@@ -96,7 +123,7 @@ interface MachineSnapshotBase<
       TTag,
       TOutput
     >,
-    testValue: ToTestStateValue<TStateValue>
+    partialStateValue: ToTestStateValue<TStateValue>
   ) => boolean;
 
   /**