+++ packages/core/src/actions/assign.ts
@@ -54,8 +54,8 @@ function resolveAssign(
       actionArgs.event,
       spawnedChildren
     ),
-    self: actorScope?.self,
-    system: actorScope?.system
+    self: actorScope.self,
+    system: actorScope.system
   };
   let partialUpdate: Record<string, unknown> = {};
   if (typeof assignment === 'function') {
@@ -99,7 +99,34 @@ export interface AssignAction<
 /**
  * Updates the current context of the machine.
  *
- * @param assignment An object that represents the partial context to update.
+ * @param assignment An object that represents the partial context to update, or a
+ * function that returns an object that represents the partial context to update.
+ * 
+ * @example
+  ```ts
+  import { createMachine, assign } from 'xstate';
+
+  const countMachine = createMachine({
+    context: {
+      count: 0,
+      message: ''
+    },
+    on: {
+      inc: {
+        actions: assign({
+          count: ({ context }) => context.count + 1
+        })
+      },
+      updateMessage: {
+        actions: assign(({ context, event }) => {
+          return {
+            message: event.message.trim()
+          }
+        })
+      }
+    }
+  });
+  ```
  */
 export function assign<
   TContext extends MachineContext,