+++ packages/core/src/types.ts
@@ -295,7 +295,7 @@ export type Actions<
 export type StateKey = string | AnyMachineSnapshot;
 
 export interface StateValueMap {
-  [key: string]: StateValue;
+  [key: string]: StateValue | undefined;
 }
 
 /**
@@ -1079,6 +1079,9 @@ export type DelayConfig<
 > = number | DelayExpr<TContext, TExpressionEvent, TParams, TEvent>;
 
 // TODO: possibly refactor this somehow, use even a simpler type, and maybe even make `machine.options` private or something
+/**
+ * @hidden
+ */
 export interface MachineImplementationsSimplified<
   TContext extends MachineContext,
   TEvent extends EventObject,
@@ -1394,24 +1397,14 @@ export type MachineConfig<
    * The machine's own version.
    */
   version?: string;
-  types?: MachineTypes<
-    TContext,
-    TEvent,
-    TActor,
-    TAction,
-    TGuard,
-    TDelay,
-    TTag,
-    TInput,
-    TOutput,
-    TTypesMeta
-  >;
   // TODO: make it conditionally required
   output?: Mapper<TContext, DoneStateEvent, TOutput, TEvent> | TOutput;
 }) &
   (MachineContext extends TContext
     ? { context?: InitialContext<LowInfer<TContext>, TActor, TInput, TEvent> }
     : { context: InitialContext<LowInfer<TContext>, TActor, TInput, TEvent> });
+
+export type UnknownMachineConfig = MachineConfig<MachineContext, EventObject>;
 
 export interface ProvidedActor {
   src: string;
@@ -1772,12 +1765,9 @@ export interface ActorOptions<TLogic extends AnyActorLogic> {
    */
   clock?: Clock;
   /**
-   * Specifies the logger to be used for log(...) actions. Defaults to the native console.log method.
+   * Specifies the logger to be used for `log(...)` actions. Defaults to the native `console.log(...)` method.
    */
   logger?: (...args: any[]) => void;
-  /**
-   * @internal
-   */
   parent?: ActorRef<any, any>;
   /**
    * @internal
@@ -1788,14 +1778,12 @@ export interface ActorOptions<TLogic extends AnyActorLogic> {
    */
   id?: string;
   /**
-   * If `true`, states and events will be logged to Redux DevTools.
-   *
-   * Default: `false`
+   * @deprecated Use `inspect` instead.
    */
-  devTools?: boolean | DevToolsAdapter; // TODO: add enhancer options
+  devTools?: never;
 
   /**
-   * The system ID to register this actor under
+   * The system ID to register this actor under.
    */
   systemId?: string;
   /**
@@ -1824,7 +1812,7 @@ export interface ActorOptions<TLogic extends AnyActorLogic> {
   state?: Snapshot<unknown>;
 
   /**
-   * The source definition.
+   * The source actor logic.
    */
   src?: string | AnyActorLogic;
 
@@ -1942,6 +1930,11 @@ export interface Subscribable<T> extends InteropSubscribable<T> {
   ): Subscription;
 }
 
+type EventDescriptorMatches<
+  TEventType extends string,
+  TNormalizedDescriptor
+> = TEventType extends TNormalizedDescriptor ? true : false;
+
 export type ExtractEvent<
   TEvent extends EventObject,
   TDescriptor extends EventDescriptor<TEvent>
@@ -1949,7 +1942,11 @@ export type ExtractEvent<
   ? TEvent
   : NormalizeDescriptor<TDescriptor> extends infer TNormalizedDescriptor
     ? TEvent extends any
-      ? TEvent['type'] extends TNormalizedDescriptor
+      ? // true is the check type here to match both true and boolean
+        true extends EventDescriptorMatches<
+          TEvent['type'],
+          TNormalizedDescriptor
+        >
         ? TEvent
         : never
       : never
@@ -2385,21 +2382,19 @@ export type ToChildren<TActor extends ProvidedActor> =
       // or maybe even `TActor["logic"]` since it's possible to configure `{ src: string; logic: SomeConcreteLogic }`
       // TODO: consider adding `| undefined` here
       Record<string, AnyActorRef>
-    : Compute<
-        ToConcreteChildren<TActor> &
-          {
-            include: {
-              [id: string]: TActor extends any
-                ? ActorRefFrom<TActor['logic']> | undefined
-                : never;
-            };
-            exclude: {};
-          }[undefined extends TActor['id'] // if not all actors have literal string IDs then we need to create an index signature containing all possible actor types
+    : ToConcreteChildren<TActor> &
+        {
+          include: {
+            [id: string]: TActor extends any
+              ? ActorRefFrom<TActor['logic']> | undefined
+              : never;
+          };
+          exclude: {};
+        }[undefined extends TActor['id'] // if not all actors have literal string IDs then we need to create an index signature containing all possible actor types
+          ? 'include'
+          : string extends TActor['id']
             ? 'include'
-            : string extends TActor['id']
-              ? 'include'
-              : 'exclude']
-      >;
+            : 'exclude'];
 
 export type StateSchema = {
   states?: Record<string, StateSchema>;