+++ templates/react-ts/src/App.tsx
@@ -1,9 +1,17 @@
 import './App.css';
 import { feedbackMachine } from './feedbackMachine';
 import { useMachine } from '@xstate/react';
+import { createBrowserInspector } from '@statelyai/inspect';
+
+const { inspect } = createBrowserInspector({
+  // Comment out the line below to start the inspector
+  autoStart: false
+});
 
 function Feedback() {
-  const [state, send] = useMachine(feedbackMachine);
+  const [state, send] = useMachine(feedbackMachine, {
+    inspect
+  });
 
   if (state.matches('closed')) {
     return (